Cleanup and setting up for Post
